requirements.txt
flask
boto3
pycryptodome
